Inline CUDA interop stubs
